Restructuring of the project
