+++ spring-boot-hello-world/src/java0.log
@@ -1,0 +1,2 @@
+oct. 27, 2022 3:48:12 P. M. org.apache.coyote.AbstractProtocol start
+INFO: Starting ProtocolHandler ["http-nio-5000"]
