+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/controller/endpoint/AudioControllerr.java
@@ -20,9 +20,9 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import com.jalasoft.convert.controller.service.FileStorageService;
 import com.jalasoft.convert.common.logger.At18Logger;
-import com.jalasoft.convert.model.Executor;
+import com.jalasoft.convert.controller.executor.Executor;
 import com.jalasoft.convert.model.commandbuilder.AudioCommand;
-import com.jalasoft.convert.model.response.AudioUploadResponse;
+import com.jalasoft.convert.controller.response.AudioUploadResponse;
 
 /**
  * It is responsable for uploading Audios and converting them
