+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/model/coverters/DocConvert.java
@@ -1,4 +1,4 @@
-package com.jalasoft.convert.model.docconvert;
+package com.jalasoft.convert.model.coverters;
 
 import com.jalasoft.convert.model.coverters.Converter;
